fix: restore merged caplen prologues, ndn_tlv_t mirror, coverage floor, pkg python3 (#177)
- proto_gtp.c: re-add #202 caplen prologues to _gtp_extract_next_extension_
  header_type and _gtp_extract_pdu_extension_header_field (dropped in merge)
  while keeping #216's extracted_data->data NULL guards
- internal_decls.h: sync the ndn_tlv_t mirror with ndn.h — #205 widened
  node_offset/data_offset to uint32_t; the stale uint16_t layout made
  ndn_test read wrong bytes through a correct parser
- build-package.sh: install python3 in the deb/rpm build containers —
  check-package-deps.sh --verify-package needs it in-container
- tests/coverage/floor.json: re-baseline to 76.0%/9 files for the wider
  instrumented set (dicom, ndn_dissector, avl suites + #200 hash_utils
  growth); decision recorded in DECISIONS.md

+++ tools/phase0/tests/internal_decls.h
@@ -348,8 +348,8 @@ typedef struct ndn_tlv_struct {
     uint16_t type;
     uint8_t nb_octets;
     unsigned long length;
-    uint16_t node_offset;
-    uint16_t data_offset;
+    uint32_t node_offset;   /* matches ndn.h — #205 widened past uint16_t */
+    uint32_t data_offset;
     struct ndn_tlv_struct *next;
 } ndn_tlv_t;
